feat: propagate logprobs in llm layer (Inferact/vllm-frontend-rs#11)

+++ rust/src/chat/src/event.rs
@@ -147,7 +147,7 @@ pub enum ChatEvent {
         message: AssistantMessage,
         /// Number of prompt tokens actually sent to the engine after chat
         /// template rendering and tokenization.
-        prompt_token_count: u32,
+        prompt_token_count: usize,
         /// Raw cumulative output token IDs, including a terminal stop token when
         /// the engine emitted one.
         token_ids: Vec<u32>,
