+++ rust/src/chat/src/output/mod.rs
@@ -45,7 +45,7 @@ pub(crate) enum AssistantEvent {
     },
     #[subenum(ContentEvent)]
     Done {
-        prompt_token_count: u32,
+        prompt_token_count: usize,
         token_ids: Vec<u32>,
         finish_reason: Option<FinishReason>,
         stop_reason: Option<StopReason>,
