+++ rust/src/text/src/output/mod.rs
@@ -13,7 +13,7 @@ use crate::{Error, Result, TextOutputStream};
 #[derive(Debug, Clone, PartialEq)]
 pub struct CollectedTextOutput {
     pub text: String,
-    pub prompt_token_count: u32,
+    pub prompt_token_count: usize,
     pub token_ids: Vec<u32>,
     pub finish_reason: Option<FinishReason>,
     pub stop_reason: Option<StopReason>,
